requirements.txt
gspread
google-auth
anthropic
playwright
requests
sendgrid
fpdf2
